requirements.txt
Django==6.0.3
dj-database-url>=2.2,<3
django-ratelimit>=4.1,<5
gunicorn>=23,<24
psycopg[binary]>=3.1.12,<4
python-dotenv>=1.0,<2
whitenoise>=6.8,<7
